refactor of CNN into classifier/regressor/network
